Fixed Product controller

+++ app/controllers/products_controller.rb
@@ -1,4 +1,5 @@
 class ProductsController < ApplicationController
+  before_action :authenticate_user!, except: [:index, :show]
   expose(:category)
   expose(:products)
   expose(:product)
@@ -15,6 +16,10 @@ class ProductsController < ApplicationController
   end
 
   def edit
+    if current_user != product.user
+      redirect_to category_product_url(category, product)
+      flash[:error] = 'You are not allowed to edit this product.'
+    end
   end
 
   def create
@@ -29,17 +34,26 @@ class ProductsController < ApplicationController
   end
 
   def update
-    if self.product.update(product_params)
-      redirect_to category_product_url(category, product), notice: 'Product was successfully updated.'
+    if current_user == product.user
+      if self.product.update(product_params)
+        redirect_to category_product_url(category, product), notice: 'Product was successfully updated.'
+      else
+        render action: 'edit'
+      end
     else
-      render action: 'edit'
+      redirect_to category_product_url(category, product)
+      flash[:error] = 'You are not allowed to edit this product.'
     end
   end
 
   # DELETE /products/1
   def destroy
-    product.destroy
-    redirect_to category_url(product.category), notice: 'Product was successfully destroyed.'
+    if current_user == product.user
+      product.destroy
+      redirect_to category_url(product.category), notice: 'Product was successfully destroyed.'
+    else
+      render action: 'edit', error: 'You are not allowed to destroy this product.'
+    end
   end
 
   private
